+++ app-openbis-command-line/src/python/obis/dm/commands/search.py
@@ -13,10 +13,11 @@
 #   limitations under the License.
 #
 
-# from queue import Queue
-# from threading import Thread
 import concurrent.futures
 
+import pandas as pd
+
+from pybis.property_reformatter import is_of_openbis_supported_date_format
 from .openbis_command import OpenbisCommand
 from ..command_result import CommandResult
 from ..utils import cd
@@ -25,11 +26,11 @@ from ...scripts.click_util import click_echo
 
 def _dfs(objects, prop, func, func_specific):
     """Helper function that perform DFS search over children graph of objects"""
+    # TODO: improve performance of this - make it similar to _dfs_samples
     with concurrent.futures.ThreadPoolExecutor(
             max_workers=5) as pool_simple, concurrent.futures.ThreadPoolExecutor(
-            max_workers=20) as pool_full:
-        stack = [getattr(openbis_obj, prop) for openbis_obj in
-                 objects]  # datasets and samples provide children in different formats
+        max_workers=20) as pool_full:
+        stack = [openbis_obj[prop] for openbis_obj in objects]  # datasets and samples provide children in different formats
         visited = set()
         stack.reverse()
         output = []
@@ -53,6 +54,90 @@ def _dfs(objects, prop, func, func_specific):
     return output
 
 
+def _dfs_samples(data_base, prop, func):
+    """Helper function that perform DFS search over children graph of objects"""
+    output = data_base
+    ids = [x['children'] for x in data_base if x['children']]
+    ids = [x[prop][prop] for x in flatten(ids)]
+    visited = set([x[prop][prop] for x in data_base])
+    while ids:
+        data = func(ids)
+        data = list(data.values())
+        output += data
+        ids = []
+        children = []
+        for obj in data:
+            key = obj[prop][prop]
+            children += [x[prop][prop] for x in obj['children']]
+            if key not in visited:
+                visited.add(key)
+        for child in children:
+            if child not in visited:
+                ids += [child]
+    return output
+
+
+def flatten(matrix):
+    flat_list = []
+    for row in matrix:
+        flat_list += row
+    return flat_list
+
+
+def _check_date(sign, date1, date2):
+    if is_of_openbis_supported_date_format(date1) and is_of_openbis_supported_date_format(date2):
+        timestamp1 = pd.to_datetime(date1)
+        timestamp2 = pd.to_datetime(date2)
+        if sign == "=":
+            return timestamp2 == timestamp1
+        elif sign == ">":
+            return timestamp2 > timestamp1
+        elif sign == "<":
+            return timestamp2 < timestamp1
+        raise ValueError(f"Unknown sign {sign}")
+    else:
+        raise ValueError("Dates are not in a supported formats!")
+
+
+def _filter_dataset(dataset, filters):
+    if filters.get("space", None) is not None:
+        space = filters["space"]
+        if dataset.sample is not None and dataset.sample.space.code != space:
+            return False
+        if dataset.experiment is not None and dataset.experiment.project.space.code != space:
+            return False
+    if filters.get("type_code", None) is not None:
+        if dataset.type.code != filters["type_code"]:
+            return False
+    if filters.get("project", None) is not None:
+        project = filters["project"]
+        if dataset.sample is not None and dataset.sample.project.code != project:
+            return False
+        if dataset.experiment is not None and dataset.experiment.project.code != project:
+            return False
+    if filters.get("experiment", None) is not None:
+        if dataset.experiment is not None and dataset.experiment.code != filters["experiment"]:
+            return False
+    if filters.get("property_code", None) is not None:
+        prop_code = filters["property_code"]
+        prop_value = filters["property_value"]
+        if dataset.props is not None and dataset.props[prop_code.lower()] != prop_value:
+            return False
+    if filters.get("registration_date", None) is not None:
+        registration_date = filters["registration_date"]
+        sign = "="
+        if registration_date[0] in [">", "<", "="]:
+            sign, registration_date = registration_date[0], registration_date[1:]
+        return _check_date(sign, registration_date, dataset.registrationDate)
+    if filters.get("modification_date", None) is not None:
+        modification_date = filters["modification_date"]
+        sign = "="
+        if modification_date[0] in [">", "<", "="]:
+            sign, modification_date = modification_date[0], modification_date[1:]
+        return _check_date(sign, modification_date, dataset.modificationDate)
+    return True
+
+
 class Search(OpenbisCommand):
     """
     Command to search samples or datasets in openBIS.
@@ -70,11 +155,16 @@ class Search(OpenbisCommand):
         self.save_path = save_path
         self.load_global_config(dm)
         self.props = "*"
-        self.attrs = ["parents", "children"]
         super(Search, self).__init__(dm)
 
     def search_samples(self):
-        search_results = self._search_samples()
+        search_results = self._search_samples(raw_response=True)
+
+        search_results = self.openbis._sample_list_for_response(props=self.props,
+                                                                response=search_results,
+                                                                attrs=["parents", "children",
+                                                                       "dataSets"],
+                                                                parsed=True)
 
         click_echo(f"Objects found: {len(search_results)}")
         if self.save_path is not None:
@@ -87,29 +177,40 @@ class Search(OpenbisCommand):
         return CommandResult(returncode=0, output="Search completed.")
 
     def _get_samples_children(self, identifier):
-        return self.openbis.get_samples(identifier, attrs=["children"])
+        return self.openbis.get_samples(identifier, attrs=["children", "dataSets"])
 
-    def _search_samples(self):
+    def _get_sample_with_datasets(self, identifier):
+        return self.openbis.get_sample(identifier, withDataSetIds=True)
+
+    def _get_sample_with_datasets2(self, identifier):
+        return self.openbis.get_sample(identifier, withDataSetIds=True, raw_response=True)
+
+    def _search_samples(self, raw_response=False):
         """Helper method to search samples"""
+
+        if self.recursive:
+            raw_response = True
 
         if "object_code" in self.filters:
             results = self.openbis.get_samples(identifier=self.filters['object_code'],
-                                               attrs=self.attrs, props=self.props)
+                                               attrs=["parents", "children", "dataSets"],
+                                               raw_response=raw_response,
+                                               props=self.props)
         else:
-            args = self._get_filtering_args(self.props)
+            args = self._get_filtering_args(self.props, ["parents", "children", "dataSets"])
+            args["raw_response"] = raw_response
             results = self.openbis.get_samples(**args)
 
         if self.recursive:
             click_echo(f"Recursive search enabled. It may take time to produce results.")
-            output = _dfs(results.objects, 'identifier',
-                          self._get_samples_children,
-                          self.openbis.get_sample)  # samples provide identifiers as children
-            search_results = self.openbis._sample_list_for_response(props=self.props,
-                                                                    response=[sample.data for sample
-                                                                              in output],
-                                                                    parsed=True)
+            output2 = _dfs_samples(results['objects'], 'identifier', self._get_sample_with_datasets2)
+
+            search_results = output2
         else:
-            search_results = results
+            if raw_response:
+                search_results = results['objects']
+            else:
+                search_results = results
         return search_results
 
     def _get_datasets_children(self, permId):
@@ -120,28 +221,56 @@ class Search(OpenbisCommand):
             return CommandResult(returncode=-1,
                                  output="Configuration fileservice_url needs to be set for download.")
 
-        if self.recursive:
-            click_echo(f"Recursive search enabled. It may take time to produce results.")
-            search_results = self._search_samples()  # Look for samples recursively
-            o = []
-            for sample in search_results.objects:  # get datasets
-                o += sample.get_datasets(
-                    attrs=self.attrs, props=self.props)
-            output = _dfs(o, 'permId',  # datasets provide permIds as children
-                          self._get_datasets_children,
-                          self.openbis.get_dataset)  # look for child datasets
+        main_filters = self.filters.copy()
+
+        object_filters = {k[7:]: v for (k, v) in main_filters.items() if k.startswith('object_')}
+        dataset_filters = {k: v for (k, v) in main_filters.items() if not k.startswith('object_')}
+        if object_filters:
+            if 'id' in object_filters:
+                if object_filters['id'] is not None:
+                    object_filters['object_code'] = object_filters['id']
+                del object_filters['id']
+            self.filters = object_filters
+            search_results = self._search_samples(raw_response=True)
+            click_echo(f"Samples found: {len(search_results)}")
+
+            datasets = [x["dataSets"] for x in search_results]
+            datasets = flatten(datasets)
+            datasets = [x['permId']['permId'] for x in datasets]
+            datasets = self.openbis.get_dataset(permIds=datasets)
+
+            filtered_datasets = []
+            for dataset in datasets:
+                if _filter_dataset(dataset, dataset_filters):
+                    filtered_datasets += [dataset]
+
             datasets = self.openbis._dataset_list_for_response(props=self.props,
-                                                               response=[dataset.data for dataset
-                                                                         in output],
+                                                               response=[x.data for x in
+                                                                         filtered_datasets],
                                                                parsed=True)
         else:
-            if "object_code" in self.filters:
-                results = self.openbis.get_sample(self.filters['object_code']).get_datasets(
-                    attrs=self.attrs, props=self.props)
+            if self.recursive:
+                search_results = self._search_samples()  # Look for samples recursively
+                o = []
+                for sample in search_results.objects:  # get datasets
+                    o += sample.get_datasets(
+                        attrs=["parents", "children"], props=self.props)
+                output = _dfs(o, 'permId',  # datasets provide permIds as children
+                              self._get_datasets_children,
+                              self.openbis.get_dataset)  # look for child datasets
+                datasets = self.openbis._dataset_list_for_response(props=self.props,
+                                                                   response=[dataset.data for
+                                                                             dataset
+                                                                             in output],
+                                                                   parsed=True)
             else:
-                args = self._get_filtering_args(self.props)
-                results = self.openbis.get_datasets(**args)
-            datasets = results
+                if "dataset_id" in self.filters:
+                    results = self.openbis.get_sample(self.filters['dataset_id']).get_datasets(
+                        attrs=["parents", "children"], props=self.props)
+                else:
+                    args = self._get_filtering_args(self.props, ["parents", "children"])
+                    results = self.openbis.get_datasets(**args)
+                datasets = results
 
         click_echo(f"Data sets found: {len(datasets)}")
         if self.save_path is not None:
@@ -153,7 +282,7 @@ class Search(OpenbisCommand):
 
         return CommandResult(returncode=0, output="Search completed.")
 
-    def _get_filtering_args(self, props):
+    def _get_filtering_args(self, props, attrs):
         where = None
         if self.filters['property_code'] is not None and self.filters['property_value'] is not None:
             where = {
@@ -163,10 +292,10 @@ class Search(OpenbisCommand):
         args = dict(space=self.filters['space'],
                     project=self.filters['project'],
                     # Not Supported with Project Samples disabled
-                    experiment=self.filters['experiment'],
+                    experiment=self.filters['collection'],
                     type=self.filters['type_code'],
                     where=where,
-                    attrs=self.attrs,
+                    attrs=attrs,
                     props=props)
 
         if self.filters['registration_date'] is not None: