+++ app-openbis-command-line/src/python/CHANGELOG.md
@@ -3,6 +3,8 @@
 * Added filtering by object in object and data_set search commands
 * Added recursive search to object and data_set search commands
 * Updated documentation regarding authentication
+* Added dataset ids to sample search results
+* changed pybis dependency to version == 1.36.0
 
 # New in version 0.4.1